requirements.txt
torch
torchvision
roma
evo
gradio
matplotlib
tqdm
opencv-python
scipy
einops
trimesh
tensorboard
pyglet<2
huggingface-hub[torch]>=0.22
gdown
imageio
cython
icecream
plyfile
open3d
spaces>=0.29.3
safetensors>=0.4.5
spandrel>=0.4.0
tqdm>=4.66.5
scikit-video>=1.1.11
diffusers==0.32.2
transformers>=4.44.0
accelerate>=0.34.2
opencv-python>=4.10.0.84
sentencepiece>=0.2.0
numpy==1.26.0
gradio>=4.44.0
imageio>=2.34.2
imageio-ffmpeg>=0.5.1
openai>=1.45.0
moviepy>=1.0.3
pillow==9.5.0
denku==0.0.51
controlnet-aux==0.0.9
wandb
peft
decord
protobuf
